Support for variable expansion in ini files.
Mustache-like `{{VAR}}` variables can be embedded in ini files and expanded at
runtime. This is useful for creating graph templates where there is only
a small amount of difference from on invocation to another such as when
graphing the numerous interfaces on a router.

Hat tip to Matt Pascoe for identifying the need.