Issue 13: split string on commas
When the function takes multiple args and a single arg is given and this arg is
a string, try splitting the arg on commas to get a list of args. If that list
has more than one element, strip whitespace from all the elements and use this
new list as the call argument list.

Also fixed a problem where strings may become double quoted if the caller has
tried to be helpful and pre-quoted the string.