Issue #15: Smarter arg splitting
Replaced simplistic `explode()` implementation with a more proper parser which
understands quoted strings and backslash escapes. A backslash before a comma
will make the comma come through as a literal with the backslash removed.
Single and double quotes can be used to surround arguments or commas as well.
The quotes will be stripped from the output. To include literal quotes in an
argument you can either surround them with the alternate quote character or
escape them with a backslash.

See the unit tests in Graphite_Graph_CallSpecTest for examples.