Issue #20: ini syntax for grouping series
Added ability to construct a series that uses the expanded versions of
multiple other series as it's basis. The derived series will have a comma
separated series list as it's starting point. Additional functions such as
`group()` can then be applied to the list of series.

By default the source series will be graphed as well as the derived series. If
you don't want to render the source series they can be marked `:is
= abstract`.

Syntax:

    [one]
    :is = abstract

    [two]
    :is = abstract

    [group]
    :series[] = one
    :series[] = two
    ; or :series = "one, two"